pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Dodomanjero <dogucanyegin81@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-binance = "1.0.19"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
